Updated tiles to include ON AIR indication
Original code included a "PS" read request to the radio to keep the connection alive. Repeated every 7 seconds. This tweak to the code changes the request to a query of the RX busy state - and I use that repeated query to change the display screen to a simple "ON AIR" indication. 
Future updates may include an ON-AIR timer as well as changes to the ON AIR font.